2023/05/28
중간 저장.
해야할 일
- 마법사 E스킬 시점에 따라가기.
- 툴팁창 오류 수정.

+++ 01Script/LSM_SC/ActorScript/PlayerMinionSC/LSM_Player_Tormented.cs
@@ -212,6 +212,6 @@ public class LSM_Player_Tormented : LSM_PlayerBase
 
         this.actorHealth.Atk =(short)((short)dummy_o[1] + add[1]);
 
-        photonView.RPC("SpawnSetting_RPC", RpcTarget.All, (short)dummy_o[0] + add[0], (short)dummy_o[0] + add[0], pname, (int)t);
+        photonView.RPC("SpawnSetting_RPC", RpcTarget.All, (short)((short)dummy_o[0] + add[0]), (short)((short)dummy_o[0] + add[0]), pname, (int)t);
     }
 }
